fix(dunst): Stopped using static values on Volume/Power
So the problem arises when things like the power-manager have multiple
responsibilities or multiple kinds of notifications like:
* Brightness indicator (with progress)
* Power Supply information (battery or on charge)
So now a static value doesn't do this justice and is also misleading. So
reverting back the changes with the risk of seeing the applet name in the notif.
Maybe we'll find a better way for this next time - since we do have the ability
to filter on message content as well.

Ref: https://dunst-project.org/documentation/#RULES