ft(term): Switched konsole for kitty
* Initial setup
* Theme selected
* Font selected
* Initial keyboard mapping setup

`Konsole` was started to slug. When compared to `kitty`, the one thing that `kitty`
doesn't provide is a running text search - which is provided only on `konsole`.
Other than that konsole is pretty sluggish and difficult to maintain as far as
configurations are concerned. `Kitty` on the other hand works well with i3wm but
does not provide the same flashy text search. What it does provide as an
alternative is the ability to search over a pager like everyone else - but hot
keys the last command that was run over `Ctrl+Shift+G`. Will use this in the
and see if we can make do with it.