fix(rofi): Address bug w.r.t executables with hyphens/uscores
There are certain rules in place to allow such exectuable names. I've been as
generic as possible:
* needs to start with alphanumeric characters
* can have hyphen(-) or underscore(_) in executable name
* needs to be followed by a man section surrounded by paranthesis like `([0-9]+)`

There are certain executables that have (seemingly legitimate) sections that are not just
numbers but alphanumeric in nature. For example `c_rehash (1ssl)`. In those cases we're basically
attempting to fetch the first section of the manpage. The manpage with section directive still works
with this as expected i.e `man c_rehash(1ssl)` gives the same results as `man c_rehash` but
something to note.

This can be of course built on in the future.