rft(polybar): Tryout top layout
The main motivation behind this is actually the `xwindow` module
used in polybar. At times *there is some confusion as to which
windown has focus*, and usually being the caveman that I am, I rapidly
try to switch focus to something else and back again if my focus was already
in it. This is only because I only use the bottom bar very lazily despite
having very active elements in it that gives me information in realtime.
By virtue of having the bar at the top, the caveman instinct in me should
scan the top first(like it does in web browsers).

Ape alone weak.